refactor: aggregator内部ロジックをAggregationContextクラスに抽出
バケツリレーされていたconn/weightCol等の共有引数をクラスのフィールドに集約し、
各メソッドの引数を削減。mkCellヘルパーも追加。外部APIは変更なし。
