+++ src/lib/aggregator.ts
@@ -1,6 +1,7 @@
 /** DuckDB SQL集計ロジック */
 
 import type * as duckdb from "@duckdb/duckdb-wasm";
+import { AggregationContext } from "./AggregationContext";
 
 // --- 集計結果の型定義 ---
 
@@ -43,6 +44,32 @@ export interface Query {
   cross_cols: QuestionDef[];
 }
 
+// --- SQL ユーティリティ（AggregationContext からも利用） ---
+
+export function esc(name: string): string {
+  return name.replace(/"/g, '""');
+}
+
+export function weightExpr(weightCol: string): string {
+  return weightCol
+    ? `SUM(TRY_CAST("${esc(weightCol)}" AS DOUBLE))`
+    : `COUNT(*)::DOUBLE`;
+}
+
+/** MA列の重み付きカウント式 */
+export function maWeightedCountExpr(maCol: string, weightCol: string): string {
+  return weightCol
+    ? `SUM(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN TRY_CAST("${esc(weightCol)}" AS DOUBLE) ELSE 0 END)`
+    : `COUNT(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN 1 END)::DOUBLE`;
+}
+
+// --- クロスヘッダーキャッシュ ---
+
+export type CrossHeaderCache = Map<
+  string,
+  { headers: Array<{ label: string; n: number }>; crossValues: string[] }
+>;
+
 /** 集計のエントリポイント。conn上のsurveyビューに対して全設問を集計する */
 export async function aggregate(
   conn: duckdb.AsyncDuckDBConnection,
@@ -50,23 +77,19 @@ export async function aggregate(
 ): Promise<AggResult[]> {
   const totalN = await computeTotalN(conn, payload.weight_col);
   const crossCols = payload.cross_cols ?? [];
-  const results: AggResult[] = [];
-
-  // クロス軸ヘッダーを事前に1回だけ取得してキャッシュ
   const crossHeaderCache =
     crossCols.length > 0
       ? await fetchCrossHeaders(conn, crossCols, payload.weight_col)
       : (new Map() as CrossHeaderCache);
 
+  const ctx = new AggregationContext(conn, payload.weight_col, totalN, crossCols, crossHeaderCache);
+
+  const results: AggResult[] = [];
   for (const q of payload.questions) {
     if (q.type === "SA") {
-      results.push(
-        await aggregateSA(conn, q.column, totalN, payload.weight_col, crossCols, crossHeaderCache)
-      );
+      results.push(await ctx.aggregateSA(q.column));
     } else {
-      results.push(
-        await aggregateMA(conn, q.prefix, q.columns, totalN, payload.weight_col, crossCols, crossHeaderCache)
-      );
+      results.push(await ctx.aggregateMA(q.prefix, q.columns));
     }
   }
 
@@ -74,23 +97,6 @@ export async function aggregate(
 }
 
 // --- 内部ユーティリティ ---
-
-function esc(name: string): string {
-  return name.replace(/"/g, '""');
-}
-
-function weightExpr(weightCol: string): string {
-  return weightCol
-    ? `SUM(TRY_CAST("${esc(weightCol)}" AS DOUBLE))`
-    : `COUNT(*)::DOUBLE`;
-}
-
-/** MA列の重み付きカウント式 */
-function maWeightedCountExpr(maCol: string, weightCol: string): string {
-  return weightCol
-    ? `SUM(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN TRY_CAST("${esc(weightCol)}" AS DOUBLE) ELSE 0 END)`
-    : `COUNT(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN 1 END)::DOUBLE`;
-}
 
 async function computeTotalN(
   conn: duckdb.AsyncDuckDBConnection,
@@ -102,11 +108,6 @@ async function computeTotalN(
   const result = await conn.query(sql);
   return Number(result.toArray()[0].n);
 }
-
-type CrossHeaderCache = Map<
-  string,
-  { headers: Array<{ label: string; n: number }>; crossValues: string[] }
->;
 
 async function fetchCrossHeaders(
   conn: duckdb.AsyncDuckDBConnection,
@@ -151,312 +152,4 @@ async function fetchCrossHeaders(
     }
   }
   return cache;
-}
-
-async function aggregateSA(
-  conn: duckdb.AsyncDuckDBConnection,
-  col: string,
-  totalN: number,
-  weightCol: string,
-  crossCols: QuestionDef[],
-  crossHeaderCache: CrossHeaderCache
-): Promise<AggResult> {
-  const sql = `
-    SELECT
-      "${esc(col)}" AS label,
-      ${weightExpr(weightCol)} AS cnt
-    FROM survey
-    WHERE "${esc(col)}" IS NOT NULL
-      AND "${esc(col)}" != ''
-    GROUP BY "${esc(col)}"
-    ORDER BY
-      TRY_CAST("${esc(col)}" AS DOUBLE) NULLS LAST,
-      "${esc(col)}" ASC
-  `;
-
-  const arrowResult = await conn.query(sql);
-  const rowArr = arrowResult.toArray();
-
-  // GT セル
-  const cells: Cell[] = rowArr.map((r) => {
-    const count = Number(r.cnt);
-    return {
-      main: String(r.label),
-      sub: "GT",
-      n: totalN,
-      count,
-      pct: totalN > 0 ? (count / totalN) * 100 : 0,
-    };
-  });
-
-  // クロスセル
-  if (crossCols.length > 0) {
-    const mainValues = rowArr.map((r) => String(r.label));
-    for (const crossQ of crossCols) {
-      const cached = crossHeaderCache.get(questionKey(crossQ))!;
-      if (crossQ.type === "SA") {
-        const crossCells = await buildCrossCellsSA(
-          conn, col, mainValues, crossQ.column, weightCol, cached
-        );
-        cells.push(...crossCells);
-      } else {
-        const crossCells = await buildCrossCellsMA(
-          conn, col, mainValues, crossQ.columns, weightCol, cached
-        );
-        cells.push(...crossCells);
-      }
-    }
-  }
-
-  return { question: col, type: "SA", cells };
-}
-
-/** SA主軸 × SA軸クロスセル生成 */
-async function buildCrossCellsSA(
-  conn: duckdb.AsyncDuckDBConnection,
-  mainCol: string,
-  mainValues: string[],
-  crossCol: string,
-  weightCol: string,
-  cached: { headers: Array<{ label: string; n: number }>; crossValues: string[] }
-): Promise<Cell[]> {
-  const { headers, crossValues } = cached;
-
-  const cellSQL = `
-    SELECT
-      "${esc(mainCol)}" AS mv,
-      "${esc(crossCol)}" AS sv,
-      ${weightExpr(weightCol)} AS cnt
-    FROM survey
-    WHERE "${esc(mainCol)}" IS NOT NULL
-      AND "${esc(mainCol)}" != ''
-      AND "${esc(crossCol)}" IS NOT NULL
-      AND "${esc(crossCol)}" != ''
-    GROUP BY "${esc(mainCol)}", "${esc(crossCol)}"
-  `;
-
-  const cellResult = await conn.query(cellSQL);
-  const cellMap = new Map<string, number>();
-  for (const r of cellResult.toArray()) {
-    cellMap.set(`${r.mv}\0${r.sv}`, Number(r.cnt));
-  }
-
-  const cells: Cell[] = [];
-  for (const mv of mainValues) {
-    for (let i = 0; i < crossValues.length; i++) {
-      const sv = crossValues[i];
-      const count = cellMap.get(`${mv}\0${sv}`) ?? 0;
-      const crossN = headers[i].n;
-      cells.push({
-        main: mv,
-        sub: sv,
-        n: crossN,
-        count,
-        pct: crossN > 0 ? (count / crossN) * 100 : 0,
-      });
-    }
-  }
-
-  return cells;
-}
-
-/** SA主軸 × MA軸クロスセル生成 */
-async function buildCrossCellsMA(
-  conn: duckdb.AsyncDuckDBConnection,
-  mainCol: string,
-  mainValues: string[],
-  maCols: string[],
-  weightCol: string,
-  cached: { headers: Array<{ label: string; n: number }>; crossValues: string[] }
-): Promise<Cell[]> {
-  const { headers } = cached;
-
-  // 各MA itemごとの、主軸値別カウントを一括取得
-  const selectClauses = maCols.map((maCol, i) => {
-    const expr = weightCol
-      ? `SUM(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN TRY_CAST("${esc(weightCol)}" AS DOUBLE) ELSE 0 END)`
-      : `COUNT(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN 1 END)::DOUBLE`;
-    return `${expr} AS c${i}`;
-  });
-
-  const sql = `
-    SELECT
-      "${esc(mainCol)}" AS mv,
-      ${selectClauses.join(", ")}
-    FROM survey
-    WHERE "${esc(mainCol)}" IS NOT NULL
-      AND "${esc(mainCol)}" != ''
-    GROUP BY "${esc(mainCol)}"
-  `;
-
-  const result = await conn.query(sql);
-  const rowMap = new Map<string, Record<string, number>>();
-  for (const r of result.toArray()) {
-    const counts: Record<string, number> = {};
-    maCols.forEach((_, i) => { counts[`c${i}`] = Number(r[`c${i}`] ?? 0); });
-    rowMap.set(String(r.mv), counts);
-  }
-
-  const cells: Cell[] = [];
-  for (const mv of mainValues) {
-    const counts = rowMap.get(mv);
-    for (let i = 0; i < maCols.length; i++) {
-      const count = counts?.[`c${i}`] ?? 0;
-      const crossN = headers[i].n;
-      cells.push({
-        main: mv,
-        sub: maCols[i],
-        n: crossN,
-        count,
-        pct: crossN > 0 ? (count / crossN) * 100 : 0,
-      });
-    }
-  }
-
-  return cells;
-}
-
-async function aggregateMA(
-  conn: duckdb.AsyncDuckDBConnection,
-  prefix: string,
-  cols: string[],
-  totalN: number,
-  weightCol: string,
-  crossCols: QuestionDef[],
-  crossHeaderCache: CrossHeaderCache
-): Promise<AggResult> {
-  const selectClauses = cols.map((col, i) => {
-    return `${maWeightedCountExpr(col, weightCol)} AS c${i}`;
-  });
-
-  const sql = `SELECT ${selectClauses.join(", ")} FROM survey`;
-  const result = await conn.query(sql);
-  const row = result.toArray()[0];
-
-  const cells: Cell[] = cols.map((col, i) => {
-    const count = Number(row[`c${i}`] ?? 0);
-    return {
-      main: col,
-      sub: "GT",
-      n: totalN,
-      count,
-      pct: totalN > 0 ? (count / totalN) * 100 : 0,
-    };
-  });
-
-  // クロスセル
-  if (crossCols.length > 0) {
-    for (const crossQ of crossCols) {
-      const cached = crossHeaderCache.get(questionKey(crossQ))!;
-      if (crossQ.type === "SA") {
-        const crossCells = await buildMACrossCellsSA(
-          conn, cols, crossQ.column, weightCol, cached
-        );
-        cells.push(...crossCells);
-      } else {
-        const crossCells = await buildMACrossCellsMA(
-          conn, cols, crossQ.columns, weightCol, cached
-        );
-        cells.push(...crossCells);
-      }
-    }
-  }
-
-  return { question: prefix, type: "MA", cells };
-}
-
-/** MA主軸 × SA軸クロスセル生成 */
-async function buildMACrossCellsSA(
-  conn: duckdb.AsyncDuckDBConnection,
-  maCols: string[],
-  crossCol: string,
-  weightCol: string,
-  cached: { headers: Array<{ label: string; n: number }>; crossValues: string[] }
-): Promise<Cell[]> {
-  const { headers, crossValues } = cached;
-
-  // クロス軸値別に各MAアイテムのカウントを取得
-  const selectClauses = maCols.map((col, i) => {
-    return `${maWeightedCountExpr(col, weightCol)} AS c${i}`;
-  });
-
-  const sql = `
-    SELECT
-      "${esc(crossCol)}" AS cv,
-      ${selectClauses.join(", ")}
-    FROM survey
-    WHERE "${esc(crossCol)}" IS NOT NULL
-      AND "${esc(crossCol)}" != ''
-    GROUP BY "${esc(crossCol)}"
-  `;
-
-  const result = await conn.query(sql);
-  const rowMap = new Map<string, number[]>();
-  for (const r of result.toArray()) {
-    const counts = maCols.map((_, i) => Number(r[`c${i}`] ?? 0));
-    rowMap.set(String(r.cv), counts);
-  }
-
-  const cells: Cell[] = [];
-  for (const maCol of maCols) {
-    const maIdx = maCols.indexOf(maCol);
-    for (let i = 0; i < crossValues.length; i++) {
-      const sv = crossValues[i];
-      const counts = rowMap.get(sv);
-      const count = counts?.[maIdx] ?? 0;
-      const crossN = headers[i].n;
-      cells.push({
-        main: maCol,
-        sub: sv,
-        n: crossN,
-        count,
-        pct: crossN > 0 ? (count / crossN) * 100 : 0,
-      });
-    }
-  }
-
-  return cells;
-}
-
-/** MA主軸 × MA軸クロスセル生成 */
-async function buildMACrossCellsMA(
-  conn: duckdb.AsyncDuckDBConnection,
-  rowMaCols: string[],
-  crossMaCols: string[],
-  weightCol: string,
-  cached: { headers: Array<{ label: string; n: number }>; crossValues: string[] }
-): Promise<Cell[]> {
-  const { headers } = cached;
-
-  // 行MA × 列MAの交差カウントを一括取得
-  const selectClauses: string[] = [];
-  for (let r = 0; r < rowMaCols.length; r++) {
-    for (let c = 0; c < crossMaCols.length; c++) {
-      const expr = weightCol
-        ? `SUM(CASE WHEN "${esc(rowMaCols[r])}" IN ('1','true') AND "${esc(crossMaCols[c])}" IN ('1','true') THEN TRY_CAST("${esc(weightCol)}" AS DOUBLE) ELSE 0 END)`
-        : `COUNT(CASE WHEN "${esc(rowMaCols[r])}" IN ('1','true') AND "${esc(crossMaCols[c])}" IN ('1','true') THEN 1 END)::DOUBLE`;
-      selectClauses.push(`${expr} AS r${r}c${c}`);
-    }
-  }
-
-  const sql = `SELECT ${selectClauses.join(", ")} FROM survey`;
-  const result = await conn.query(sql);
-  const row = result.toArray()[0];
-
-  const cells: Cell[] = [];
-  for (let r = 0; r < rowMaCols.length; r++) {
-    for (let c = 0; c < crossMaCols.length; c++) {
-      const count = Number(row[`r${r}c${c}`] ?? 0);
-      const crossN = headers[c].n;
-      cells.push({
-        main: rowMaCols[r],
-        sub: crossMaCols[c],
-        n: crossN,
-        count,
-        pct: crossN > 0 ? (count / crossN) * 100 : 0,
-      });
-    }
-  }
-
-  return cells;
 }
